refactor(resources): Move reactive components outside of page.tsx
